Introducing Service layer. Created UsersService.
