+++ src/test/java/integrationTests/UserTests.java
@@ -3,6 +3,7 @@ package integrationTests;
 import org.testng.annotations.BeforeClass;
 import org.testng.annotations.Test;
 import users.UsersClient;
+import users.UsersService;
 import users.create.CreateUserRequestBody;
 import users.create.response.CreateUserResponse;
 
@@ -10,11 +11,11 @@ import java.util.UUID;
 
 
 public class UserTests {
-    private UsersClient usersClient;
+    private UsersService usersService;
 
     @BeforeClass
     public void beforeClass() {
-        usersClient = new UsersClient();
+        usersService = new UsersService();
     }
 
     @Test
@@ -23,9 +24,9 @@ public class UserTests {
         CreateUserRequestBody requestBody = new CreateUserRequestBody.Builder().build();
 
         // 2. Act
-        int id = usersClient.createUser(requestBody).getData().getId();
+        int id = usersService.createUser(requestBody).getData().getId();
 
         // 3. Assert
-        usersClient.getUser(id).assertUser(requestBody);
+        usersService.getUser(id).assertUser(requestBody);
     }
 }
