+++ src/test/java/users/UsersClient.java
@@ -3,30 +3,10 @@ package users;
 import io.restassured.http.ContentType;
 import io.restassured.response.Response;
 import users.create.CreateUserRequestBody;
-import users.create.response.CreateUserErrorResponse;
-import users.create.response.CreateUserResponse;
-import users.get.GetUsersResponse;
-import users.getAll.GetAllUsersResponse;
 
 import static io.restassured.RestAssured.given;
 
 public class UsersClient {
-    public CreateUserResponse createUser(CreateUserRequestBody body) {
-        Response response = create(body);
-        CreateUserResponse createUserResponse = response.as(CreateUserResponse.class);
-        createUserResponse.setStatusCode(response.statusCode());
-
-        return createUserResponse;
-    }
-
-    public CreateUserErrorResponse createUserExpectingError(CreateUserRequestBody body) {
-        Response response = create(body);
-        CreateUserErrorResponse errorResponse = response.as(CreateUserErrorResponse.class);
-        errorResponse.setStatusCode(response.statusCode());
-
-        return errorResponse;
-    }
-
 
     public static Response create(CreateUserRequestBody body) {
         Response response =
@@ -44,23 +24,20 @@ public class UsersClient {
         return response;
     }
 
-    public static GetAllUsersResponse getAllUsers() {
-        Response response = given()
+    public static Response getAll() {
+        Response response =
+                given()
                 .when()
                     .get("https://gorest.co.in/public/v1/users");
 
-        response.then()
-                .log().body();
+        response
+                .then()
+                    .log().body();
 
-        int statusCode = response.statusCode();
-
-        GetAllUsersResponse getAllUsersResponse = response.as(GetAllUsersResponse.class);
-        getAllUsersResponse.setStatucCode(statusCode);
-
-        return getAllUsersResponse;
+        return response;
     }
 
-    public GetUsersResponse getUser(int id) {
+    public static Response get(int id) {
         Response response =
                 given()
                         .header("Authorization", "Bearer 457a143dbd57bef631313eb0e02b0777a8d268c3ae52c43ea34aa4581d7e0ed5")
@@ -71,11 +48,6 @@ public class UsersClient {
                 .then()
                     .log().body();
 
-        int statusCode = response.statusCode();
-
-        GetUsersResponse getUsersResponse = response.as(GetUsersResponse.class);
-        getUsersResponse.setStatusCode(statusCode);
-
-        return getUsersResponse;
+        return response;
     }
 }
